Trying to consolidate a single source of truth for the package versions...

+++ scripts-postbuild/build.rs
@@ -25,23 +25,23 @@ fn main() {
     // This directry can then easily be include-ed in C++ projects
 
     let target = env::var("TARGET").unwrap();
-    println!("cargo:warning=target: {}", target);
+    println!("cargo:info=target: {}", target);
 
     let profile = env::var("PROFILE").unwrap();
-    println!("cargo:warning=profile: {}", profile);
+    println!("cargo:info=profile: {}", profile);
 
     let project_dir = project_root();
-    println!("cargo:warning=project_dir: {}", project_dir.display());
+    println!("cargo:info=project_dir: {}", project_dir.display());
 
     // We will be targeting the 'jcan' subdirectory - this is hard-coded.
     let manifest_dir = project_dir;
 
     // Delete the project-level out directory
     let out_dir = Path::new(&manifest_dir).join("out");
-    println!("cargo:warning=project_out_dir: {}", out_dir.display());
+    println!("cargo:info=project_out_dir: {}", out_dir.display());
 
     let out_dir = Path::new(&manifest_dir).join("out").join(&profile).join(&target).join("jcan");
-    println!("cargo:warning=out_dir: {}", out_dir.display());
+    println!("cargo:info=out_dir: {}", out_dir.display());
 
     // If is_debug, copy the debug library, otherwise copy the release library
     // If iscrossbuild, another 'target' subdir is added
@@ -54,7 +54,15 @@ fn main() {
     // Create the out directory
     fs::create_dir_all(&out_dir).unwrap();
 
-    // Copy the library to the out directory
+    // Check if the library exists
+    if !lib_build_path.exists() {
+        println!("Library not found at {}", lib_build_path.display());
+        println!("Quitting");
+        // Quit
+        return;
+    }
+
+    // Copy the library to the out directory 
     fs::copy(&lib_build_path, &out_dir.join("libjcan.a")).unwrap();
 
     // Copy the cxxbridge generated headers to the out directory
